--- src/main/java/Task.java
@@ -1,67 +1,0 @@
-package main.java;
-
-// Base Task class
-public abstract class Task {
-    String description;
-    boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public void markAsDone() {
-        isDone = true;
-    }
-
-    public void markAsNotDone() {
-        isDone = false;
-    }
-
-    @Override
-    public abstract String toString();
-}
-
-// ToDo subclass
-class ToDo extends Task {
-    public ToDo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + (isDone ? "[X] " : "[ ] ") + description;
-    }
-}
-
-// Deadline subclass
-class Deadline extends Task {
-    String by;
-
-    public Deadline(String description, String by) {
-        super(description);
-        this.by = by;
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + (isDone ? "[X] " : "[ ] ") + description + " (by: " + by + ")";
-    }
-}
-
-// Event subclass
-class Event extends Task {
-    String start;
-    String end;
-
-    public Event(String description, String start, String end) {
-        super(description);
-        this.start = start;
-        this.end = end;
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + (isDone ? "[X] " : "[ ] ") + description + " (from: " + start + " to: " + end + ")";
-    }
-}
